requirements.txt
cycler==0.10.0
dill==0.3.4
groot-trees==0.0.15
gurobipy==9.1.2
joblib==1.0.1
kiwisolver==1.3.2
llvmlite==0.37.0
matplotlib==3.4.3
numba==0.54.0
numpy==1.20.0
pandas==1.3.2
Pillow==9.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-sat==0.1.7.dev10
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
threadpoolctl==2.2.0
tqdm==4.62.2
roct==0.0.3
